requirements.txt
# Opencode Local Audio Toolkit Requirements

# Common
numpy
soundfile

# Speech-to-Text (STT)
faster-whisper
sounddevice
scipy

# Text-to-Speech (TTS)
kokoro-onnx
onnxruntime-gpu

# NVIDIA CUDA 12 Runtime (Windows/Linux)
# These satisfy onnxruntime-gpu and faster-whisper (ctranslate2) dependencies
nvidia-cublas-cu12
nvidia-cudnn-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-nvjitlink-cu12
nvidia-cufft-cu12
nvidia-nvtx-cu12
nvidia-cuda-runtime-cu12
nvidia-cuda-nvrtc-cu12
